refactor(LC017): split projection analysis helpers
- separate query-chain, usage, and syntax-based property analysis
- keep analyzer behavior unchanged while reducing local complexity
